requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-appconf==1.0.3
django-autoslug==1.9.4
django-contact-form==1.7
django-filer==1.5.0
django-forms-bootstrap==3.1.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mptt==0.10.0
django-pagedown==1.0.6
django-pipeline==1.6.14
django-polymorphic==2.0.2
django-tinymce==2.8.0
easy-thumbnails==2.6
idna==2.8
markdown2==2.3.7
Pillow==6.0.0
psycopg2-binary==2.8.2
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.25.3
